Cache sigmas for better performance
When you use gridgen-c from the command line, it optionally writes
the sigmas values to a file, so that when you tweak the parameters,
the computationally intensive task of computing the sigmsa doesn't
have to be repeated. This commit caches the sigmas as a property of
the class andf sets `nx`, `ny`, and `focus` to properties as well.
Now you can create a gridgen object, visualize it, tweak the shape
and focus, and then regenerate the grid with significantly improved
performance.